gradebook/requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.1
boto>=2.4
requests>=2.18
numpy>=1.1
beautifulsoup4>=4.3.0
pycrypto>=2.6.1
pymongo==3.6.0
